pyproject.toml
[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.poetry]
name = "lczero-easy"
version = "0.1.0"
description = "LeelaChessZero made easy."
license = "MIT"
authors = [
    "Yoann Poupart <yoann.poupart@ens-lyon.org>",
]

[[tool.poetry.source]]
name = "PyPI"
priority = "default"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
python-chess = "^1.999"
torch = "^2.1.0"
onnx2pytorch = "^0.4.1"
matplotlib = "^3.8.2"
torchviz = "^0.0.2"
zennit = "^0.5.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
pytest = "*"
pytest-cov = "*"

[tool.poetry.group.api]
optional = true

[tool.poetry.group.api.dependencies]
python-multipart = "^0.0.6"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}

[tool.poetry.group.demo]
optional = true

[tool.poetry.group.demo.dependencies]
gradio = "^4.8.0"
